+++ src/main/java/com/voucher/vouchermanagement/manager/VoucherManager.java
@@ -6,8 +6,8 @@ import com.voucher.vouchermanagement.manager.io.VoucherManagerOutput;
 import com.voucher.vouchermanagement.model.customer.Customer;
 import com.voucher.vouchermanagement.model.voucher.Voucher;
 import com.voucher.vouchermanagement.model.voucher.VoucherType;
-import com.voucher.vouchermanagement.service.BlacklistService;
-import com.voucher.vouchermanagement.service.VoucherService;
+import com.voucher.vouchermanagement.service.blacklist.BlacklistService;
+import com.voucher.vouchermanagement.service.voucher.VoucherService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
