+++ src/main/java/com/voucher/vouchermanagement/service/blacklist/BlacklistService.java
@@ -1,4 +1,4 @@
-package com.voucher.vouchermanagement.service;
+package com.voucher.vouchermanagement.service.blacklist;
 
 import com.voucher.vouchermanagement.model.customer.Customer;
 import com.voucher.vouchermanagement.repository.blacklist.BlacklistFileRepository;
